+++ src/main/java/com/project/Aperetif/Services/interfaces/EventDetailsService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface EventDetailsService {
     int saveEvents(EventDetails event);
 
-    EventDetails getEventById(Long id);
+    EventDetails getEventById(Integer id);
 
     List<EventDetails> findAll();
